Rename testing_utils to testing_constants
