+++ pallets/currency/src/testing_constants.rs
